move getWarningCount to custom dir
